python/requirements.txt
pyserial==3.5
numpy==1.24.3
matplotlib==3.7.1
